requirements.txt
amqp
anyjson
backports-abc
backports.ssl-match-hostname
beautifulsoup4
billiard
boto
celery
certifi
cffi
couchbase
cryptography
Django
django-filter
django-grappelli
djangorestframework
djangorestframework-filters
docutils
drf-extensions
enum34
flower
gunicorn
idna
ipaddress
isodate
jsonfield
kombu
Markdown
ndg-httpsclient
numpy
oauthlib
psycopg2
pyasn1
pycparser
pyOpenSSL
python-dateutil
pytz
PyYAML
requests
requests-oauthlib
scikit-learn
scipy
singledispatch
six
subprocess32
tornado
ujson
urllib3
